third_party/WebKit/LayoutTests/fast/dom/shadow/dynamically-added-deep-combinator-and-shadow-pseudo-element-expected.txt

CONSOLE WARNING: /deep/ combinator is deprecated. See https://www.chromestatus.com/features/6750456638341120 for more details.
CONSOLE WARNING: ::shadow pseudo-element is deprecated. See https://www.chromestatus.com/features/6750456638341120 for more details.
Should have blue background.
PASS backgroundColorOf(getNodeInTreeOfTrees("host/target1")) is "rgb(0, 128, 0)"
PASS backgroundColorOf(getNodeInTreeOfTrees("host/target2")) is "rgb(0, 128, 0)"
PASS backgroundColorOf(getNodeInTreeOfTrees("host/target3")) is "rgb(0, 0, 255)"
PASS backgroundColorOf(getNodeInTreeOfTrees("host/target4")) is "rgb(0, 0, 255)"
PASS backgroundColorOf(getNodeInTreeOfTrees("target5")) is "rgb(0, 0, 255)"
PASS successfullyParsed is true

TEST COMPLETE
